extract Convertor class
