+++ src/main/java/redis/clients/johm/collections/RedisList.java
@@ -11,7 +11,7 @@ import redis.clients.johm.Indexed;
 import redis.clients.johm.JOhm;
 import redis.clients.johm.JOhmUtils;
 import redis.clients.johm.Nest;
-import redis.clients.johm.JOhmUtils.Convertor;
+import redis.clients.johm.Convertor;
 import redis.clients.johm.JOhmUtils.JOhmCollectionDataType;
 
 /**
