+++ src/main/java/redis/clients/johm/collections/RedisMap.java
@@ -12,7 +12,7 @@ import redis.clients.johm.Indexed;
 import redis.clients.johm.JOhm;
 import redis.clients.johm.JOhmUtils;
 import redis.clients.johm.Nest;
-import redis.clients.johm.JOhmUtils.Convertor;
+import redis.clients.johm.Convertor;
 import redis.clients.johm.JOhmUtils.JOhmCollectionDataType;
 
 /**
@@ -123,7 +123,7 @@ public class RedisMap<K, V> implements Map<K, V> {
         for (String key : nest.cat(JOhmUtils.getId(owner)).cat(field.getName())
                 .hkeys()) {
             if (johmKeyType == JOhmCollectionDataType.PRIMITIVE) {
-                keys.add((K) JOhmUtils.Convertor.string2object(keyClazz, key));
+                keys.add((K) Convertor.string2object(keyClazz, key));
             } else if (johmKeyType == JOhmCollectionDataType.MODEL) {
                 keys.add(JOhm.<K> get(keyClazz, Integer.parseInt(key)));
             }
@@ -204,13 +204,13 @@ public class RedisMap<K, V> implements Map<K, V> {
         for (Map.Entry<String, String> entry : savedHash.entrySet()) {
             if (johmKeyType == JOhmCollectionDataType.PRIMITIVE
                     && johmValueType == JOhmCollectionDataType.PRIMITIVE) {
-                savedKey = (K) JOhmUtils.Convertor.string2object(keyClazz, entry
+                savedKey = (K) Convertor.string2object(keyClazz, entry
                         .getKey());
-                savedValue = (V) JOhmUtils.Convertor.string2object(valueClazz, entry
+                savedValue = (V) Convertor.string2object(valueClazz, entry
                         .getValue());
             } else if (johmKeyType == JOhmCollectionDataType.PRIMITIVE
                     && johmValueType == JOhmCollectionDataType.MODEL) {
-                savedKey = (K) JOhmUtils.Convertor.string2object(keyClazz, entry
+                savedKey = (K) Convertor.string2object(keyClazz, entry
                         .getKey());
                 savedValue = JOhm.<V> get(valueClazz, Integer.parseInt(entry
                         .getValue()));
@@ -218,7 +218,7 @@ public class RedisMap<K, V> implements Map<K, V> {
                     && johmValueType == JOhmCollectionDataType.PRIMITIVE) {
                 savedKey = JOhm.<K> get(keyClazz, Integer.parseInt(entry
                         .getKey()));
-                savedValue = (V) JOhmUtils.Convertor.string2object(valueClazz, entry
+                savedValue = (V) Convertor.string2object(valueClazz, entry
                         .getValue());
             } else if (johmKeyType == JOhmCollectionDataType.MODEL
                     && johmValueType == JOhmCollectionDataType.MODEL) {
